custom-elements/instantiating-custom-elements/custom-element-type-local-name.html

<!DOCTYPE html>
<html>
<head>
<title>Custom element type is the local name of the custom element</title>
<meta name="author" title="Sergey G. Grekhov" href="mailto:sgrekhov@unipro.ru">
<meta name="assert" content="The custom element type can be given to a custom element at the time of its instantation as the local name of the custom element">
<link rel="help" href="http://www.w3.org/TR/custom-elements/#instantiating-custom-elements">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src='../testcommon.js'></script>
<link rel="stylesheet" href="/resources/testharness.css">
</head>
<body>
<div id="log"></div>
<script type="text/javascript">

test(function() {
    var doc = newHTMLDocument();
    
    var ElementConstructor = doc.register('x-a-1');
    
    var customElement = new ElementConstructor();
    
    assert_equals(customElement.constructor.prototype.constructor.name, 
            customElement.localName, 'Wrong custom element type');
        
}, 'Custom element type is the local name of the custom element. ' + 
    'Custom element created in JavaScript');


test(function() {
    var doc = newHTMLDocument();
    
    doc.register('x-a-2');
    
    doc.body.innerHTML = '<x-a-2 id="ce2"></x-a-2>';
    
    var customElement = doc.querySelector('#ce2');
    
    assert_equals(customElement.constructor.prototype.constructor.name, customElement.localName, 
        'Wrong custom element type');
        
}, 'Custom element type is the local name of the custom element.' + 
    'Custom element created in HTML');
    

testInIFrame('../resources/x-element.html', function(context) {
    var doc = context.iframes[0].contentDocument;
    
    doc.register('x-element');

    var xelement = doc.querySelector('#xelement');
    
    assert_equals(xelement.constructor.prototype.constructor.name, xelement.localName, 
        'Wrong custom element type');

}, 'Custom element type is the local name of the custom element. ' + 
    'Custom element created in HTML. Document has browsing context');


</script>
</body>
</html>
